Fix: the if condition about the length string

+++ file_io/1-create_file.c
@@ -16,7 +16,7 @@ int _strlen(char *s)
 
 /**
  * create_file - function that creates a file
- * @filemane: file
+ * @filemane: name of the file to create
  * @text_content: string to write into the file
  * Return: 1 or -1 if failed
  */
@@ -24,7 +24,7 @@ int _strlen(char *s)
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	ssize_t len;
+	int len;
 
 	if (filename == NULL)
 	{
